# Conflicts:
#	src/1. Layers/1.2 Services/Well.Services/Contracts/IExceptionEventService.cs
#	src/1. Layers/1.2 Services/Well.Services/ExceptionEventService.cs
#	src/1. Layers/1.3 Domain/Well.Domain/Well.Domain.csproj

+++ src/1. Layers/1.3 Domain/Well.Domain/JobDetailDamage.cs
@@ -18,7 +18,7 @@
         public int DamageActionId { get; set; }
 
         [XmlIgnore]
-        public decimal Qty { get; set; }
+        public int Qty { get; set; }
 
         [XmlElement("Qty")]
         public string QtyXml
